// src/modules/sdk-flex-card/elements/view/layout/flex/types/flex.ts
import { AnyObject } from "../../../../../../../shared/types/common/AnyObject";
import { ElementType } from "../../../../core/types/element";

export type FlexType = {
  elementType: "grid";
  element: "flex";
  version: number;
  childrens?: ElementType[];
  styles?: AnyObject;
  props?: FlexPropsType;
};

export type FlexPropsType = {
  vertical?: boolean;
  wrap?: "nowrap" | "wrap" | "wrap-reverse";
  justify?:
    | "start"
    | "center"
    | "space-between"
    | "space-around"
    | "space-evenly";
  align?: "stretch" | "center" | "start" | "end";
  gap?: "small" | "middle" | "large";
};